添加iobuffer

+++ common/iointerface.go
@@ -1,6 +1,6 @@
 package common
 
 type TSMUXIO interface {
-	Read([]byte, int) error
-	Write([]byte, int) error
+	Read([]byte, uint64) error
+	Write([]byte, uint64) error
 }
